refactor(server): refactoring utils to classes and types folders
