+++ server/src/utils/types/GlobalContext.ts
@@ -1,5 +1,5 @@
 import { ExpressContext } from 'apollo-server-express';
-import Member from '../entities/Member';
+import Member from '../../entities/Member';
 
 export type GlobalContext = ExpressContext & {
 	user: Member | null;
